SWDEV-230497 cherry-pick into amd-mainline-open
This is recommit of f13d10bf621967869121313e654fa3e874a980f6, which was
reverted due to build failure of MIOpen. Now that build failure of
MIOpen was fixed by https://reviews.llvm.org/D78655, recommit this.

https://reviews.llvm.org/D77954 caused a regression about ambiguity of new operator
in file scope.

This patch recovered the previous behavior for comparison without a caller.

This is a workaround. For real fix we need D71227

https://reviews.llvm.org/D78970

Change-Id: I76de9b7ed2624c9706b97c0f81264a0c78ece496