[clang-offload-bundler] Add option -fail-on-missing-bundles
There are out-of-tree tools using clang-offload-bundler to extract
bundles from bundled files. When a bundle is not in the bundled
file, clang-offload-bundler is expected to emit an error message
and return non-zero value. However currently clang-offload-bundler
silently generates empty file for the missing bundles.

Since OpenMP toolchain expects the current behavior, an option
-fail-on-missing-bundles is added to let clang-offload-bundler
fail when a bundle is missing when unbundling.

Differential Revision: https://reviews.llvm.org/D93068

Change-Id: Ibe0e2c0eeb317c2413703d913e96639cca5f3cf8