AMDGPU: Add second emergency slot for SGPR to vmem for large frames
In a future change, we will sometimes use a VGPR offset for doing
spills to memory, in which case we need 2 free VGPRs to do the SGPR
spill. In most cases we could spill the VGPR along with the SGPR being
spilled, but we don't have any free lanes for SGPR_1024 in wave32 so
we could still potentially need a second scavenging slot.

(cherry picked from commit d6fdbbcace0b51c0096c5dbab6afb6449da21524)
Change-Id: I395c34c74e50ffa4db23d66798d6cf335a74c213