Re-commit "[LLD] Improve reporting unresolved symbols in shared libraries"
Fixes: SWDEV-306719

This patch reverts the workaround 3c0f24db6b12fdb19b41ff078d9519115cc6c012
and cherry-pick the real fix ce25eb1.

[ELF] Do not report undefined weak references in shared libraries

This fixes an issue introduced in D101996.

A weak reference in a shared library could be incorrectly reported if
there is another library that has a strong reference to the same symbol.

Differential Revision: https://reviews.llvm.org/D115041

Change-Id: I8f6782f13917ed3a7e940dfd28afe1277bc492b5