[MachineSink] Allow sinking of constant or ignorable physreg uses
For AMDGPU, any use of the physical register EXEC prevents sinking even if it is not a real physical register read. Add check to see if a physical
register use can be ignored for sinking.

Also perform same constant and ignorable physical register check when considering sinking in loops.

regenerated tests for:
  LLVM :: CodeGen/AMDGPU/GlobalISel/sdiv.i64.ll
  LLVM :: CodeGen/AMDGPU/GlobalISel/srem.i64.ll
  LLVM :: CodeGen/AMDGPU/GlobalISel/udiv.i64.ll
  LLVM :: CodeGen/AMDGPU/GlobalISel/urem.i64.ll
  LLVM :: CodeGen/AMDGPU/frame-index-elimination.ll
  LLVM :: CodeGen/AMDGPU/mul24-pass-ordering.ll
  LLVM :: CodeGen/AMDGPU/skip-if-dead.ll

https://reviews.llvm.org/D116053

Change-Id: Iba9c782d169b7e0a9649f86dc1fc7d866d0cdb98