[HeterogeneousDwarf] Initial support for globals at -O0
This implements only the "optimized" case where the DIFragment for a
DIGlobalVariable refers only to a single global-variable DIFragment.

The general case for DIOpArg requires DIE-reference arguments, and will
require some work along the same lines as the existing DW_OP_convert
support.

This should support most globals at -O0, and it gracefully drops debug
information for cases it cannot handle.

Note that the patch is unaware that no symbols are defined for LDS
globals or locals, nor is it aware that LDS locals are not true globals
even though they are represented via LLVM global values. This
technically regresses the debug information as we will go from
describing all LDS variables as "optimized out" to "always at LDS
address 0x0", which is not correct. This will be addressed in future
patches.

Change-Id: I4d4a76548e6becf2a5738060bd482004b01d7d1f