[HeterogeneousDwarf] Fix asserts at -O0
This patch addresses some cases where rather than gracefully
drop debug information we assert, even when compiling at -O0

This effectively constitutes a workaround, and the underlying issues
still need to be investigated and corrected.

Both changes relate to cases where the debug information is not (or
cannot) be directly connected to an `alloca` referrer, due to e.g. ABI
constraints on parameters.

The first change is to SDagIsel, where intention was always to drop the
debug info (at least for now) in some cases, hoever this was never
correctly signaled to the caller who checks for `nullptr`. Return
`nullptr` in these cases so the debug info is dropped.

The second change is to RegAllocFast, where we were never explicitly
dropping debug info in cases we know we cannot handle, leading to
assertions later. In reviewing this code, I believe we were also not
respecting the requirement that debug info have no effect on codegen.
The true fix for this will have to be accompanied by work in the DWARF
backend to support DIOpReferrer more generally, but for now RegAllocFast
is just taught to cleanly drop the debug info it cannot handle.

Change-Id: I94c5784f96493835904ba7f4e278bb6ea0975df2