[HeterogeneousDwarf][SelectionDAG] Fix use-after-return
This use-after-return only seems to trigger observable failures in some
exceptional circumstances, so there is no practical test to accompany
the patch.

Fix amdgpu-debug-info-variable-expression-heterogeneous-dwarf.cl

In some builds clang will default to eliding non-global value names in
IR identifiers, which confused this test. Explicitly request that the
identifier names not be elided.

Change-Id: I7f1198fb1c70a86fae9d60bd388fb994d1af9e62