[OpenMP] Fix the static library build for DeviceRTL in libomptarget
This issue prevented the static lib libomptarget.devicertl.a to
build, since the commented part of CMakeLists.txt was originally
building this static library. When uncommented, this initiated the
compilation of various files within DeviceRTL/src. These files
had various compilation errors which are fixed in this patch and
now the libomptarget.devicertl.a is building successfully.

Change-Id: If5f2a7a4c162ab88fea1132e9eb960262e7ff620