Handle continue in no-loop code generation.
If a continue is encountered directly in the loop for which a no-loop
kernel is generated, add an unconditional branch to the end of the
kernel. If a continue is encountered in a nested loop, normal code
generation should proceed. Implemented this functionality without
changing the Emit-APIs by keeping metadata for the No-Loop statement.

Change-Id: I3aa626f218140edcdcc0d1d53d22ffc240fe41e2