[llvm][Attributor] - Comment out call to getAssumedSimplified.
This is causing crashes in opt where dead constants
are trying to be removed and do not exist in the map.
This may be related to when we link in the
libomptarget-amdgcn-gfxABC.bc. Trunk uses
mlink-builtin-bitcode during device compilation and has
no issues. ASO uses llvm-link to combine bitcode and then
opt fails. This is problem is seen on non-gfx90a cards."

Change-Id: I338aca41c0307105a9857dfdce97a4fb697f5909