SWDEV-328212 don't busy wait
With just a 'yield', the OS will potentially release the CPU, but will
instantly require it, resulting in the CPU just contributing to global
warming as it spins waiting for jobs to complete using 100% of the cpu.

A quick and dirty solution is to put in a tiny sleep (1 millisecond)
in the thread waiting other threads to complete. This is what is
implemented here. With 4 compiling threads and one monitoring thread,
this saves about 20% of the cpu.

A better solution would pobably involve condition variables so the
thread was only woken up when something had actually completed. Since
there is an URGENT need to make this system "jobserver" aware (see
https://www.gnu.org/software/make/manual/html_node/Job-Slots.html ) to
avoid the current thrashing we observe, this better solution can come
as part of that work.

Change-Id: Ic7afbc6179333f7e7d53392ee91a0c396b34dd48
Signed-off-by: Icarus Sparry <icarus.sparry@amd.com>