[clang][driver] - Fix bug where the suffix logic is incorrect.
If an archive such as libclang_rt.builtins-x86_64.a
is encountered the current logic will split the
string based on the first '.', which results in
an archive being improperly turned into an object
file. This patch properly gets the suffix in such
scenarios and avoids an issue we are seeing
in clang build select link during qmcpack build.
This patch was needed for normal compilation as
well as -save-temps.

Change-Id: Ibf0f01e7e910d53c19b7c4a3b0e3cc459f7bf724