[AMDGPU] Preserve the dst register of wwm-copy operations
When SGPRs are spilled into virtual VGPR lanes by letting
regalloc to efficiently allocate the VGPRs. Allocator might
possibly split the liverange of such VGPRs by introducing
an intermediate copy or by spilling to a superclass which
is guaranteed to be a whole wave copy with the exec mask
manipulation around it.

In addition, the entire lanes of the destination register
should be preserved at function prolog/epilog as its inactive
lanes can be clobbered. The patch ensures that by adding
them into wwm spills.

Fixes: SWDEV-391411, SWDEV-391413

Change-Id: I610d12376850eca9a67569675ae50ab78d82fbae