[OpenMP] - Updates to ClangOffloadWrapper and nextgen amdgpu plugin.
The nextgen plugin expects an offload binary for
the image instead of the old ELF format. This is one
difference between the NEW/OLD driver. This patch takes
pieces from the NEW driver's ClangLinkerWrapper and ports
them to the OLD driver's ClangOffloadWrapper. Currently,
there was no way to extract metadata (arch id for example)
from the image as the nextgen plugin needs to receive an
offload binary.

The runtime change is to support systems that have multiple
gpu archs from the same vendor. For example, a system with
a gfx908 and gfx90a has runtime issues validating the binary
even when compiling with --offload-arch=gfx908,gfx90a as
the rule is if the image does not match ALL of the AMD
devices then validation will not occur. Now, if ANY device
supports that image then it is valid.

However, there is still a problem. If the user compiles for
both archs and uses device(1) on a target pragma the runtime
loads the wrong image and then crashes. We need to track
which image works on which device to solve this.

Change-Id: I5c6f713f7a3dbaaaf073351a64779df0d5dc6ce1