[OpenMP] [AMDGPU] Fixed behavior of env-var OMPX_FORCE_SYNC_REGIONS.
This env-var is used to force synchronous target regions. The default is
to use an asynchronous implementation. Unlike in the legacy plugin, it
is not enough to force a synchronous codepath from libomptarget for data
submit and retrieve for nextgen AMDGPU plugin. Hence, we now read the
env-var explicitly in the plugin as well and if set, the submit and
retrieve will follow the synchronous codepaths.

The existing isSynchronous interface in the async wrapper has
been removed since that could return true for the asynchronous case
as well during data movement in the GenericGlobalHandler.

No change is required in kernel launch since that already checks for
synchronous behavior correctly. A comment has been added to that effect.

Change-Id: I4b4701fb94bdbf165eb5b815ce1d486c17d71297