[OpenMP] Fixes observed hang from async handler
We observed hangs instead of program crashes after the initial
implementation of the lazy HSA queue creation. It appears that the HSA
async handler was causing this behavior.

This patch remvoes the async handler, and instead uses the nextgen
plugins stream slots to schedule an action once the slot has finished.
This uses the same overall design but a different mechanism to decrement
a queues busy count.

Change-Id: I584a62ef81144c4a8d4881c01d9c3e7e5fdbe79a