[AMDGPU] Add ASM and MC updates for preloading kernargs
Cherry-pick 69447d6afe702dee6bc1dbe17f4915789ad6889d

Add assembler directives for preloading kernel arguments that correspond
to new fields in the kernel descriptor for the length and offset of
arguments that will be placed in SGPRs prior to kernel launch. Alignment
of the arguments in SGPRs is equivalent to the kernarg segment when
accessed via the kernarg_segment_ptr. Kernarg SGPRs are allocated
directly after other user SGPRs.

Reviewed By: arsenm

Differential Revision: https://reviews.llvm.org/D159459

Change-Id: I6f9c0f10c9cb80766ac16c1eab6249dfe1049ee0