[AMDGPU] Use inreg for hint to preload kernel arguments
Cherry-pick 60a227c464a19a00a76d19b5bc75e0e4d5c89873

This patch is the first in a series that adds support for pre-loading
kernel arguments into SGPRs. The command-line argument
'amdgpu-kernarg-preload-count' is used to specify the number of
arguments sequentially from the first that we should attempt to preload,
the default is 0.

Reviewed By: arsenm

Differential Revision: https://reviews.llvm.org/D156852

Change-Id: I45e298fab6691421b296ccdaf250fefa8065c6e2