[Comgr] Refactor nested kernel behavior into separate test
OpenCL allows a kernel function to be called from another kernel function,
and this behavior is featured in the primary OpenCL files used in the Comgr
tests. However, this behavior is atypical and rarely-used.

In this patch, we refactor this behavior into a separate test,
and restructure the primary tests to not call nested kernels. This will allow
us to more quickly identify issues specifically with nested kernels without
failing a large array of tests.

Change-Id: I1772db85b770977012798940dd2cba5fc2cdf8ac