[HeterogeneousDwarf] Keep track of GlobalVariables DIFragments using WeakVH
In DwarfDebug::beginModule, the compiler stores for each fragment its
associated GlobalVariable in a DenseMap.
However, nothing prevents the GlobalVariable from being removed during a later pass,
invalidating the reference in the map.
In the attached test case, AMDGPULowerLDS removes LDS variables without any use,
which resulted in an use-after-free access.

The WeakVH works as a reference to the GlobalVariable, however, when the
global is erased a callback sets the WeakVH to invalid.

Change-Id: If1627f548441f07ab32454ece2beefb518ab2aa2