CycleInfo: Fix splitCriticalEdge (#68584)
There are cases when cycle that contains both Pred and Succ is not
found: Pred is in a smaller cycle that is contained in a larger cycle
that also contains Succ, or Pred and Succ are in disjointed innermost
cycles but there is a larger cycle that contains both.
Add efficient helper function that finds innermost cycle that contains
both Pred and Succ if it exists.

Change-Id: I9f96f8fa3c7dda5702205191df8e04c0b685957d