[AMDGPU][SILowerSGPRSpills] Add fallback path to legacy sgpr spill handling
Added the commandline switch `amdgpu-legacy-sgpr-spill-lowering` to enable
the legacy spill lowering method. This flag is turned off by default as the
new spill handling method is preferred in the compiler.

This switch is only temporary and can be used when the default spill method
produces incorrect codegen. This flag will be removed when the new spill
handling path is fully functional and handles all corner cases.

Change-Id: Ib9a1e470c3e8cdc1a595894af625b19952012c97