[CUDA][HIP] Fix std::is_invocable
Currently std::is_invocable does not work for CUDA/HIP since
its implementation requires checking whether a function
is invocable in the context of a synthesized host function.

In general, to make <type_traits> work with CUDA/HIP, the
template functions need to be defined as
so that they are available in both host and device contexts.

Fixes: https://github.com/llvm/llvm-project/issues/69956

Fixes: SWDEV-428314