[OpenMP][MI300A][libomptarget] Add option to make map constructs prefault pages on target GPU.
Preafulting pages allows to execute applications built with xnack- in unified_shared_memory mode.
A translation for each mapped page is added to the GPU page table and there is no need to run the XNACK algorithm when a page is referenced by a GPU thread for the first time.

Change-Id: Ib3939510ecadc05e8a8f9a109d6f9d28eef3811b