[Comgr] Shift postopt linking option to OpenCL compilation path
Previously, -mlink-builtin-bitcode-postopt was applied whenever
the device libraries were included. However, this inadvertently
included the option for the newer *COMPILE_TO_RELOCATABLE action.

This inclusion introduced a performance regression in some hipRTC
applications, which is currently the only API using
*COMPILE_TO_RELOCATABLE.

To address this, in this patch we refactor the postopt linking
option to only apply to *COMPILE_TO_BITCODE, its originally intended
action.

Change-Id: I2b3c2936ff093ae587ca9695138fa723418850c4