Test Only: Do not merge: [HIP] Fix __clang_hip_cmath.hip for ambiguity
If there is a type T which can be converted to both float and double etc but itself
is not specialized for __numeric_type, and it is called for math functions eg. fma,
it will cause ambiguity with test function of __numeric_type.

Since test is not template, this error is not bypassed by SFINAE. This is a design
flaw of __numeric_type. This patch fixes clang wrapper header to use SFINAE to
avoid such ambiguity.

Fixes: SWDEV-461604

Fixes: https://github.com/llvm/llvm-project/issues/101239

Change-Id: Ia7942cdee54450ec30e726a6bbcb93f6a6f6fe83