[AMDGPU] Fix SDWA commute
SDWA insts miss reverse opcode, which causes them to be treated
as commutable with default reverse opcode i.e. their own opcode.
As a result, SWDA F16 sub A, B and Sub B, A are commonized by
machine CSE. The correct bavior is to commonize sub A, B and subrev B, A
instead of sub B, A. This issues cause failures in rocFFT.

Another issue is that src0_sel and src1_sel are not swapped
when SDWA insts are commutated.

Change-Id: I20e19bfd403542d7f865535fd463dd6c1e514f8e