[AMDGPU] Move kernarg preload logic to AMDGPU Attributor
Besides the changes listed below everything works the same as it did
when this code was in AMDGPULowerKernelArguments.

There is a refactoring of the free user SGPR tracking to make it
simplified and more accurate. We don't actually care which SGPRs hold
which arguments before ISel so specific tracking of the number of free
registers is removed. In one case this leads to one extra argument being
preloaded in a test. ISel correctly identifies this opportunity even
when the IR pass previously missed it. Even though inreg is meant to act
as a hint the coupling between the attribute and whether an argument is
actually preloaded should be equivalent now, although ISel always makes
the final determination.

Since we are no longer handling this in AMDGPULowerKernelArguments that
pass must rely on the inreg attribute to determine whether to leave
arguments as is. This leads to some test changes.

This lowering is moved out of the llc pipeline which requires test
updates.

Cloned function declarations are removed when kernel signatures are
modified to preload hidden arguments.

Change-Id: I11aaa432e528bc973084242673644b1c1ef1d9b1