MI-400: Revert const phys regs folding (#3677)
Reason:

I does not work. The assumption was: "With current codegen this only
affects src_flat_scratch_base_lo/hi." It is not. It actually affects
src_private_base on gfx9, and it does not work because Pierre have
defined src_private_base_hi which does not exist, and he has defined it
with the same encoding as src_private_base. Actually all 4 apperture
registers are prone to this.

This also has nothing to do with MI400 as it affects gfx9 codegen.

Feel free to reintroduce it in upstream.