Cherry-pick AMDGPU: Change DS classes to use RegisterOperand parameters (#156580)
Cherry-pick commit c5a8841 in main into amd-gfx13. Unify VGPROp and
VGPRSrc.

Start stripping out the uses of getLdStRegisterOperand. This
added a confusing level of indirection where the class at the
definition point was not the actual class used. This was also
pulling in the AV class usage for targets where it isn't
relevant. This was also inflexible for special cases.

Also fixes using default arguments which only served to wrap the
class argument in a RegisterOperand.

This should be done for all the memory instructions.