MI-400: Multicast load bug workaround for MI450 A0. (#4143)
As described in SWDEV-556203, there is a multicast cluster load bug in
MI450 A0 when the multicast mask (M0) is not 0. This bug can lead to
gfxip crash or data leakage between processes. To workaround it, the
compiler can set M0 to 0 before the cluster load, ignoring the actual
input of the multicast mask. This workaround means we will
unicast-return the result immediately to the requester, avoiding the
hardware issue.

Note: Do not upstream this patch, remove it when B0 is available.

Fixes: SWDEV-556203