[NFC] Fix for llvm/test/CodeGen/AMDGPU/bundle-idx-loadstore.mir test. (#4184)
Although the test is marked as aut-generated, it contains a bunch of
manually edited lines where unstable numbers changed to masks. So in the
following regeneratation of the test in 9a2e41d these edits were
overwritten.

The root cause is for inline metadata without slots printer outputs
objects' pointers. It's not obvious how the printer should be fixed, so
I've just restored masks in the places.

Fixes SWDEV-554952.