AMDGPU/GFX13: Bundle two-address instructions with V_{LOAD,STORE}_IDX (#4736)
An important part of this change is to ensure that earlyclobber and tied
constraints are applied correctly to bundled V_{LOAD,STORE}_IDX:

* When operands are tied and we bundle one of them, we must be able to
bundle both of them, with the same index + offset. Otherwise, we can
attempt to convert the relevant instruction into its 3-address form.

* When a def is earlyclobber and we bundle the corresponding
V_STORE_IDX, we must ensure that it does not alias any of the uses
(except possibly a tied use).

In order to integrate these checks somewhat more cleanly, this change
includes a fairly large refactoring of the bundling logic. Instead of
trying to apply all the checks on-the-fly, we build a list of candidate
operands for bundling and then filter that list down based on
constraints.

An additional fairly large conceptual change is that instead of sinking
all
instructions forming the bundle towards the last V_STORE_IDX, we instead
keep the "core" instruction where it is and move all V_LOAD/STORE_IDX
towards it.

That change makes some of the alias checking logic easier to follow.
Apart from that, it ought to be fairly benign. If anything, this seems
like it ought to be slightly better because the core instruction stays
where it is relative to everything else, i.e. there should be fewer
surprising changes in the instruction schedule.

v2:
- remove the code to handle tied IMPLICIT_DEF uses -- that cannot really
happen for us in useful code, so let's keep the code simple

v3:
- fix an issue with V_STORE_IDX index operand depending indirectly on
the core MI