[GFX13] Remove redundant wait on VmVsrc in expert scheduling mode (#4855)
The logic in clearRedundantVmVsrcWait() was not correct. It should be
checking if there are any pending events on VM_VSRC, not checking if
there is an actual wait on that counter.

Fix this, and rename the function simplifyVmVsrc(), and call it from
simplifyWaitcnt(&Wait) rather than determineWait(). Simplification of
VmVsrc is also treated more conservatively.

This also fixed a newly-discovered issue where a pre-existing
S_WAIT_DEPCTR waiting on VM_VSRC could have that wait removed entirely
when it should be preserved.

Update a number of tests that were failing to distinguish correct versus incorrect
when using s_wait_alu.

(cherry picked from commit be9cf22365461139ab4b19c97652977f2d6f481d)