AMDGPU: Simplify idx0 handling and private-as-VGPR allocas (#4866)
Normalize the semantics of indices for V_{LOAD,STORE}_IDX: Make them
always absolute from instruction selection. This avoids the weird shift
in semantics based on metadata during the bundle-indexed-load-store
pass.
(Previously, indices were implicitly relative to idx0 based on metadata
analysis. We want to get away from that.)

At the same time, simplify how idx0 is saved and restored:
amdgpu-idx-reg-alloc will now simply copy from idx0 at the beginning of
the function into a virtual register and use that copy throughout.

The baseline lowering for this copy is an s_getreg_b32 instruction.
Frame lowering (prolog insertion) has a heuristic that is often able to
elide this instruction in the entry block.