AMDGPU: Treat idx0 writes as a scheduling barrier (#4867)
As the comment indicates, most instructions depend implicitly on idx0 in
GFX13. However, adding implicit operands everywhere would be death by a
thousand cuts for compile time (even more than we already get due to
implicit $exec operands...). Instead, we treat instructions that set
idx0 as hard scheduling boundaries. This ought to help avoid potential
bugs when idx0 is temporarily changed from the private VGPR base.

This also means that prolog insertion no longer has to insert a generic
SCHED_BARRIER. This changes instruction schedules in the prolog (because
the scheduling boundary is now in a different place). The changes are a
bit of a mixed bag. There's nothing major there, and if anything, the
way to solve this is to give the scheduler more precise, target-specific
information. In any case, not using the blunt generic scheduling barrier
puts us on a better path.