AMDGPU/GFX13: Improve some private-in-vgpr tests (#4910)
The heuristics of AMDGPUMarkPromotablePrivate don't align with those of
AMDGPUPromoteAlloca yet, and most of these tests would fail them. Tweak
the tests to also satisfy the rules of AMDGPUPromoteAlloca to prepare
for future developments.

Increasing the size of the allocation in the subdword test prevents the
allocas from being promoted to vector.