AMDGPU/GFX13: Preserve live intervals & friends in AMDGPUPrivateObjectVGPRs (#4911)
The subsequent two-address instruction pass doesn't require these
analyses, but it can dos more powerful transforms when they are present.
So avoid invalidating them.

There are currently no affected tests. I noticed this with a
work-in-progress change that turns this pass on unconditionally, which
affected:

* llvm/test/CodeGen/AMDGPU/load-lo16.ll
* llvm/test/CodeGen/AMDGPU/mad-mix-lo.ll

These tests don't have GFX13 RUN lines yet, and I hesitate to add them
just for this -- it seems like that would increase the risk of merge
conflicts.