AMDGPU/GFX13: Allocate private-as-VGPR already in LLVM IR (#4912)
This change has two goals:

* Make the object address more easily accessible to downstream passes.
* Make it easier to make smarter allocation decisions by moving the
allocation out of instruction selection.

The test changes are mildly annoying, but they're simply down to the
allocas now being allocated in a different order. THe new order is in a
sense more natural, since it is in the order the allocas appear in the
IR.