[GFX13] Support instructions using STOREcnt completing out-of-order (#4498)
VIMAGE and ray tracing instructions that track completion using STOREcnt
can complete out-of-order with respect to each other and other
instructions using STOREcnt. Add two new event types in SIInsertWaitcnts
for these classes of instruction and associate them with STOREcnt so
that the existing out-of-order completion logic will do the right thing
when they are encountered.

This means there are now 33 event types in total, so widen the type used
for bitmasks representing sets of wait events to uint64_t. Add an
assertion to check that future event additions have not overflowed the
mask size.

This also adds a test case for RTS_RAY_SAVE, which introduces for the
first time a VGPR data dependency requirement using STOREcnt.